Move service-worker files into SW directory
