+++ src/sw/managers/sessionManager/SessionManager.ts
@@ -1,7 +1,7 @@
-import { ContextSWInterface } from "../../../models/ContextSW";
-import { PushDeviceRecord } from "../../../models/PushDeviceRecord";
-import { SessionOrigin } from "../../../models/Session";
-import { ISessionManager } from "../types";
+import { ContextSWInterface } from "../../../shared/models/ContextSW";
+import { PushDeviceRecord } from "../../../shared/models/PushDeviceRecord";
+import { ISessionManager } from "../../../shared/managers/sessionManager/types";
+import { SessionOrigin } from "../../../shared/models/Session";
 
 export class SessionManager implements ISessionManager {
   constructor(_context: ContextSWInterface) { }
